requirements.txt
aiogram
opencv-python-headless
pyserial
pytz
